Start refactoring python and robot callback creators into a helper class
