+++ packages/assistant/src/RPA/Assistant/library.py
@@ -35,10 +35,10 @@ from flet_core import Stack
 from flet_core.control_event import ControlEvent
 from flet_core.dropdown import Option
 from robot.api.deco import keyword, library
-from robot.errors import RobotError
 from robot.libraries.BuiltIn import BuiltIn, RobotNotRunningError
 from robot.utils.dotdict import DotDict
 from typing_extensions import Literal
+from RPA.Assistant.callback_runner import CallbackRunner
 
 from RPA.Assistant.flet_client import FletClient
 from RPA.Assistant.types import (
@@ -181,6 +181,7 @@ class Assistant:
         self.logger = logging.getLogger(__name__)
         os.environ["FLET_LOG_LEVEL"] = "warning"
         self._client = FletClient()
+        self._callbacks = CallbackRunner(self._client)
         self._validations: Dict[str, Callable] = {}
         self._required_fields: Set[str] = set()
         self._open_layouting: List[str] = []
@@ -600,15 +601,16 @@ class Assistant:
 
         if validation:
             if isinstance(validation, str):
-                self._validations[name] = self._create_robot_validation_wrapper(
-                    validation
-                )
+                self._validations[
+                    name
+                ] = self._callbacks._create_robot_validation_wrapper(validation)
             elif isinstance(validation, Callable):
-                self._validations[name] = self._create_python_validation_wrapper(
-                    validation
-                )
+                self._validations[
+                    name
+                ] = self._callbacks._create_python_validation_wrapper(validation)
             else:
                 raise ValueError("Invalid validation function.")
+
         if required:
             self._required_fields.add(name)
 
@@ -1087,109 +1089,6 @@ class Assistant:
         """
         self._client.update_elements()
 
-    def _create_python_function_wrapper(
-        self, function: Callable[[Any], None], *args, **kwargs
-    ) -> Callable[[], None]:
-        """wrapper code that is used to add wrapping for user functions when binding
-        them to be run by buttons
-        """
-
-        def func_wrapper() -> None:
-            return self._run_python_callback(function, *args, **kwargs)
-
-        return func_wrapper
-
-    def _create_python_validation_wrapper(
-        self, function: Callable[[Any], Optional[str]]
-    ) -> Callable[[Any], Optional[str]]:
-        """wrapper code that is used to add wrapping for user functions when binding
-        them to be run on validations buttons
-        """
-
-        def func_wrapper(value) -> Optional[str]:
-            self._run_python_callback(function, value)
-
-        return func_wrapper
-
-    def _run_python_callback(self, function, *args, **kwargs):
-        try:
-            return function(*args, **kwargs)
-
-        # This can be anything since it comes from the user function, we don't
-        # want to let the user function crash the UI
-        except Exception as err:  # pylint: disable=broad-except
-            self.logger.error(f"Error calling Python function {function}")
-            self.logger.error(err)
-        finally:
-            self._client.unlock_elements()
-            self._client.flet_update()
-
-    def _create_robot_function_wrapper(
-        self, kw_name: str, *args, **kwargs
-    ) -> Callable[[], None]:
-        """wrapper code that is used to add wrapping for user functions when binding
-        them to be run by buttons
-        """
-
-        def func_wrapper() -> None:
-            return self._run_robot_callback(kw_name, *args, **kwargs)
-
-        return func_wrapper
-
-    def _create_robot_validation_wrapper(
-        self, kw_name: str
-    ) -> Callable[[Any], Optional[str]]:
-        """wrapper code that is used to add wrapping for user functions when binding
-        them to be run on validation
-        """
-
-        def func_wrapper(value) -> Optional[str]:
-            return self._run_robot_callback(kw_name, value)
-
-        return func_wrapper
-
-    def _run_robot_callback(self, kw_name: str, *args, **kwargs):
-        try:
-            self._client.lock_elements()
-            self._client.flet_update()
-            return BuiltIn().run_keyword(kw_name, *args, **kwargs)
-        except RobotNotRunningError:
-            self.logger.error(
-                f"Robot Framework not running so cannot call keyword {kw_name}"
-            )
-        except RobotError as e:
-            self.logger.error(f"Error calling robot keyword {kw_name}")
-            self.logger.error(e)
-        # This can be anything since it comes from the user function, we don't
-        # want to let the user function crash the UI
-        except Exception as err:  # pylint: disable=broad-except
-            self.logger.error(f"Unexpected error running robot keyword {kw_name}")
-            self.logger.error(err)
-        finally:
-            self._client.unlock_elements()
-            self._client.flet_update()
-
-    def _queue_function_or_robot_keyword(
-        self, function: Union[Callable, str], *args, **kwargs
-    ):
-        """Check if function is a Python function or a Robot Keyword, and call it
-        or run it with Robot's run_keyword.
-        """
-        if self._client.pending_operation:
-            self.logger.error("Can't have more than one pending operation.")
-            return
-        self._client.lock_elements()
-        self._client.flet_update()
-
-        if isinstance(function, Callable):
-            self._client.pending_operation = self._create_python_function_wrapper(
-                function, *args, **kwargs
-            )
-        else:
-            self._client.pending_operation = self._create_robot_function_wrapper(
-                function, *args, **kwargs
-            )
-
     @keyword(tags=["dialog"])
     def add_button(
         self,
@@ -1220,7 +1119,7 @@ class Assistant:
         """
 
         def on_click(_: ControlEvent):
-            self._queue_function_or_robot_keyword(function, *args, **kwargs)
+            self._callbacks._queue_function_or_keyword(function, *args, **kwargs)
 
         button = ElevatedButton(label, on_click=on_click)
         container = Container(alignment=location.value, content=button)
@@ -1259,7 +1158,7 @@ class Assistant:
         """
 
         def on_click(_: ControlEvent):
-            self._queue_function_or_robot_keyword(function, self._get_results())
+            self._callbacks._queue_function_or_keyword(function, self._get_results())
 
         button = ElevatedButton(label, on_click=on_click)
         self._client.add_element(button)
